+++ src/test/java/nextstep/subway/acceptance/AcceptanceTest.java
@@ -1,6 +1,6 @@
 package nextstep.subway.acceptance;
 
-import nextstep.subway.utils.DatabaseCleanup;
+import nextstep.subway.testhelper.DatabaseCleanup;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
