+++ src/test/java/nextstep/subway/testhelper/DatabaseCleanup.java
@@ -1,4 +1,4 @@
-package nextstep.subway.utils;
+package nextstep.subway.testhelper;
 
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.context.annotation.Profile;
