--- src/test/java/nextstep/subway/unit/LineTest.java
@@ -1,17 +1,0 @@
-package nextstep.subway.unit;
-
-import org.junit.jupiter.api.Test;
-
-class LineTest {
-    @Test
-    void addSection() {
-    }
-
-    @Test
-    void getStations() {
-    }
-
-    @Test
-    void removeSection() {
-    }
-}
